requirements.txt
asgiref==3.7.2
certifi==2023.11.17
crispy-bootstrap5==2023.10
distlib==0.3.8
Django==5.0.1
django-bower==5.2.0
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-scheduler==0.10.1
filelock==3.13.1
icalendar==5.0.11
pillow==10.3.0 # deleted in pluto
pip==23.3.2
pipenv==2023.11.15
platformdirs==4.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
setuptools==69.0.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
virtualenv==20.25.0